working with Top10 of last entered, and filtered. tests need to be changed
