+++ react/pom.xml
@@ -112,71 +112,44 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 			</plugin>
-			
-			<plugin>
-		      <artifactId>maven-assembly-plugin</artifactId>
-		      <configuration>
-		        <archive>
-		          <manifest>
-		            <mainClass>de.booking.ReactAndSpringDataRestApplication</mainClass>
-		          </manifest>
-		        </archive>
-		        <descriptorRefs>
-		          <descriptorRef>jar-with-dependencies</descriptorRef>
-		        </descriptorRefs>
-		      </configuration>
-		    </plugin>
 
-
-			<!-- Plugins below are for testing JSX files-->
 			<plugin>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.2</version>
+				<groupId>com.github.eirslett</groupId>
+				<artifactId>frontend-maven-plugin</artifactId>
+				<version>1.2</version>
 				<configuration>
-				<source>1.8</source>
-				<target>1.8</target>
-				</configuration>
-			</plugin>
-
-			<plugin>
-				<groupId>com.github.klieber</groupId>
-				<artifactId>phantomjs-maven-plugin</artifactId>
-				<version>0.7</version>
-				<!--configuration>
-				<version>2.1.1</version>
-				</configuration-->
-			</plugin>
-			<plugin>
-				<groupId>com.github.searls</groupId>
-				<artifactId>jasmine-maven-plugin</artifactId>
-				<version>2.2</version>
-				<configuration>
-				<specRunnerTemplate>REQUIRE_JS</specRunnerTemplate>
-				<preloadSources>
-					<source>/webjars/jsx-requirejs-plugin/0.6.0/js/require.js</source>
-					<source>/spec/spec_helper.js</source>
-				</preloadSources>
-				<jsSrcDir>${project.basedir}/src/main/js</jsSrcDir>
-				<jsTestSrcDir>${project.basedir}/src/test/resources/spec</jsTestSrcDir>
-				<specExcludes>
-				  <exclude>**/spec_helper.js</exclude>
-				</specExcludes>
-				<specIncludes>
-					<include>**/*.js</include>
-					<include>**/*.jsx</include>
-				</specIncludes>
+					<installDirectory>target</installDirectory>
 				</configuration>
 				<executions>
 					<execution>
+						<id>install node and npm</id>
 						<goals>
-						<goal>test</goal>
+							<goal>install-node-and-npm</goal>
+						</goals>
+						<configuration>
+							<nodeVersion>v4.4.5</nodeVersion>
+							<npmVersion>3.9.2</npmVersion>
+						</configuration>
+					</execution>
+					<execution>
+						<id>npm install</id>
+						<goals>
+							<goal>npm</goal>
+						</goals>
+						<configuration>
+							<arguments>install</arguments>
+						</configuration>
+					</execution>
+					<execution>
+						<id>webpack build</id>
+						<goals>
+							<goal>webpack</goal>
 						</goals>
 					</execution>
 				</executions>
 			</plugin>
-
-
 		</plugins>
 	</build>
+
 
 </project>
